infracciones

+++ src/main/java/com/gamq/ambiente/serviceimplement/InfraccionServiceImpl.java
@@ -7,9 +7,11 @@ import com.gamq.ambiente.exceptions.ResourceNotFoundException;
 import com.gamq.ambiente.model.DetalleInspeccion;
 import com.gamq.ambiente.model.Infraccion;
 import com.gamq.ambiente.model.Inspeccion;
+import com.gamq.ambiente.model.TipoInfraccion;
 import com.gamq.ambiente.repository.DetalleInspeccionRepository;
 import com.gamq.ambiente.repository.InfraccionRepository;
 import com.gamq.ambiente.repository.InspeccionRepository;
+import com.gamq.ambiente.repository.TipoInfraccionRepository;
 import com.gamq.ambiente.service.InfraccionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Example;
@@ -27,6 +29,8 @@ public class InfraccionServiceImpl implements InfraccionService {
     @Autowired
     InfraccionRepository infraccionRepository;
     @Autowired
+    TipoInfraccionRepository tipoInfraccionRepository;
+    @Autowired
     InspeccionRepository inspeccionRepository;
     @Autowired
     DetalleInspeccionRepository detalleInspeccionRepository;
@@ -41,8 +45,8 @@ public class InfraccionServiceImpl implements InfraccionService {
     }
 
     @Override
-    public List<InfraccionDto> obtenerInfraccionPorFechaInfraccion(Date fechaInfraccion) {
-        List<Infraccion> infraccionList = infraccionRepository.findByFechaInfraccion(fechaInfraccion);
+    public List<InfraccionDto> obtenerInfraccionPorFecha(Date fecha) {
+        List<Infraccion> infraccionList = infraccionRepository.findByFechaInfraccion(fecha);
         return infraccionList.stream().map(infraccion -> {
             return InfraccionMapper.toInfraccionDto(infraccion);
         }).collect(Collectors.toList());
@@ -58,34 +62,37 @@ public class InfraccionServiceImpl implements InfraccionService {
 
     @Override
     public InfraccionDto crearInfraccion(InfraccionDto infraccionDto) {
-        Inspeccion inspeccion = inspeccionRepository.findByUuid(infraccionDto.getInspeccionDto().getUuid())
-                .orElseThrow(() -> new RuntimeException("Inspección no encontrada"));
-        boolean excedeLimite = evaluarResultadoInspeccion(infraccionDto.getInspeccionDto().getUuid());
-        if (excedeLimite) {
-            if (!infraccionRepository.existsByInspeccionUuidAndTipoInfraccionUuid(infraccionDto.getInspeccionDto().getUuid(),
-                    infraccionDto.getTipoInfraccionDto().getUuid())) {
-                Infraccion nuevoInfraccion = InfraccionMapper.toInfraccion(infraccionDto);
-                nuevoInfraccion.setInspeccion(inspeccion);
-                return InfraccionMapper.toInfraccionDto(infraccionRepository.save(nuevoInfraccion));
-            }
+        Optional<Inspeccion> inspeccionOptional = inspeccionRepository.findByUuid(infraccionDto.getInspeccionDto().getUuid());
+        if (inspeccionOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Inspeccion","uuid", infraccionDto.getInspeccionDto().getUuid());
         }
-        else {
-            throw new BlogAPIException("409-CONFLICT", HttpStatus.CONFLICT, "No se genera la infraccion");
+        Optional<TipoInfraccion> tipoInfraccionOptional = tipoInfraccionRepository.findByUuid(infraccionDto.getTipoInfraccionDto().getUuid());
+        if (tipoInfraccionOptional.isEmpty()) {
+            throw new ResourceNotFoundException("Tipo Infraccion","uuid", infraccionDto.getTipoInfraccionDto().getUuid());
         }
-        throw new BlogAPIException("409-CONFLICT", HttpStatus.CONFLICT, "la Infraccion ya existe");
+        Infraccion nuevoInfraccion = InfraccionMapper.toInfraccion(infraccionDto);
+        nuevoInfraccion.setInspeccion(inspeccionOptional.get());
+        nuevoInfraccion.setTipoInfraccion(tipoInfraccionOptional.get());
+        return InfraccionMapper.toInfraccionDto(infraccionRepository.save(nuevoInfraccion));
     }
 
     @Override
     public InfraccionDto actualizarInfraccion(InfraccionDto infraccionDto) {
         Optional<Infraccion> infraccionOptional = infraccionRepository.findByUuid(infraccionDto.getUuid());
         if(infraccionOptional.isPresent()) {
-           // if (!infraccionRepository.exitsInfraccionLikeDescripcion(InfraccionDto.getDescripcion().toLowerCase(), InfraccionDto.getUuid())) {
-                Infraccion updateInfraccion = InfraccionMapper.toInfraccion(infraccionDto);
-                updateInfraccion.setIdInfraccion(infraccionOptional.get().getIdInfraccion());
-                return InfraccionMapper.toInfraccionDto(infraccionRepository.save(updateInfraccion));
-           // } else {
-            //    throw new BlogAPIException("409-CONFLICT", HttpStatus.CONFLICT, "la Infraccion ya existe");
-            //}
+            Optional<Inspeccion> inspeccionOptional = inspeccionRepository.findByUuid(infraccionDto.getInspeccionDto().getUuid());
+            if (inspeccionOptional.isEmpty()) {
+                throw new ResourceNotFoundException("Inspeccion","uuid", infraccionDto.getInspeccionDto().getUuid());
+            }
+            Optional<TipoInfraccion> tipoInfraccionOptional = tipoInfraccionRepository.findByUuid(infraccionDto.getTipoInfraccionDto().getUuid());
+            if (tipoInfraccionOptional.isEmpty()) {
+                throw new ResourceNotFoundException("Tipo Infraccion","uuid", infraccionDto.getTipoInfraccionDto().getUuid());
+            }
+            Infraccion updateInfraccion = InfraccionMapper.toInfraccion(infraccionDto);
+            updateInfraccion.setIdInfraccion(infraccionOptional.get().getIdInfraccion());
+            updateInfraccion.setTipoInfraccion(tipoInfraccionOptional.get());
+            updateInfraccion.setInspeccion(inspeccionOptional.get());
+            return InfraccionMapper.toInfraccionDto(infraccionRepository.save(updateInfraccion));
         }
         throw new ResourceNotFoundException("Infraccion", "uuid",infraccionDto.getUuid());
     }
@@ -96,9 +103,9 @@ public class InfraccionServiceImpl implements InfraccionService {
         Optional<Infraccion> optionalInfraccion = infraccionRepository.findOne(Example.of(infraccionQBE));
         if(optionalInfraccion.isPresent()){
             Infraccion infraccion = optionalInfraccion.get();
-        /*    if(!Infraccion.getInfraccionInspeccionList().isEmpty()){
-                throw new BlogAPIException("400-BAD_REQUEST", HttpStatus.BAD_REQUEST, "la Infraccion ya esta siendo usado por las inspecciones");
-            }*/
+            if (infraccion.isEstadoPago()) {
+                throw new BlogAPIException("400-BAD_REQUEST", HttpStatus.BAD_REQUEST, "la Infraccion ha sido cancelada no puede eliminar");
+            }
             infraccionRepository.delete(infraccion);
             return InfraccionMapper.toInfraccionDto(infraccion);
         }
